app/pyproject.toml
[tool.poetry]
name = "poll-service-api"
version = "0.1.0"
description = ""
authors = ["Aleksandr Smelik <suntrope55@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.8.19"
fastapi="^0.111.1"
uvicorn="^0.30.3"
motor="^3.5.1"
pydantic-settings="^2.4.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-asyncio = "0.23.8"
pytest-cov = "5.0.0"
pytest-sugar = "1.0.0"
ruff = "0.6.1"
mypy = "1.11.1"
pre-commit = "3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
strict = true
disable_error_code = ["assignment", "import", "call-arg", "return", "type-arg", "var-annotated", "arg-type"]
exclude = ["venv", ".venv"]

[tool.ruff]
target-version = "py38"
line-length = 100
exclude = ["alembic"]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG001", # unused arguments in functions
]

ignore = [
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
